+++ bigquery/src/test/java/com/google/cloud/hadoop/io/bigquery/output/ForwardingBigQueryFileOutputCommitterTest.java
@@ -21,7 +21,7 @@ import static org.mockito.Mockito.when;
 
 import com.google.cloud.hadoop.fs.gcs.InMemoryGoogleHadoopFileSystem;
 import com.google.cloud.hadoop.io.bigquery.BigQueryFileFormat;
-import com.google.cloud.hadoop.util.testing.CredentialConfigurationUtil;
+import com.google.cloud.hadoop.util.testing.CredentialsConfigurationUtil;
 import com.google.common.collect.ImmutableList;
 import java.io.IOException;
 import java.util.List;
@@ -132,7 +132,7 @@ public class ForwardingBigQueryFileOutputCommitterTest {
     // Setup the configuration.
     job = Job.getInstance(InMemoryGoogleHadoopFileSystem.getSampleConfiguration());
     conf = job.getConfiguration();
-    CredentialConfigurationUtil.addTestConfigurationSettings(conf);
+    CredentialsConfigurationUtil.addTestConfigurationSettings(conf);
     BigQueryOutputConfiguration.configure(
         conf,
         QUALIFIED_TEST_TABLE_ID,