+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/testing/TestConfiguration.java
@@ -17,8 +17,7 @@ package com.google.cloud.hadoop.gcsio.testing;
 /** Access to test configurations values. */
 public abstract class TestConfiguration {
   public static final String GCS_TEST_PROJECT_ID = "GCS_TEST_PROJECT_ID";
-  public static final String GCS_TEST_SERVICE_ACCOUNT = "GCS_TEST_SERVICE_ACCOUNT";
-  public static final String GCS_TEST_PRIVATE_KEYFILE = "GCS_TEST_PRIVATE_KEYFILE";
+  public static final String GCS_TEST_JSON_KEYFILE = "GCS_TEST_JSON_KEYFILE";
 
   /** Environment-based test configuration. */
   public static class EnvironmentBasedTestConfiguration extends TestConfiguration {
@@ -28,13 +27,8 @@ public abstract class TestConfiguration {
     }
 
     @Override
-    public String getServiceAccount() {
-      return System.getenv(GCS_TEST_SERVICE_ACCOUNT);
-    }
-
-    @Override
-    public String getPrivateKeyFile() {
-      return System.getenv(GCS_TEST_PRIVATE_KEYFILE);
+    public String getServiceAccountJsonKeyFile() {
+      return System.getenv(GCS_TEST_JSON_KEYFILE);
     }
   }
 
@@ -48,7 +42,5 @@ public abstract class TestConfiguration {
 
   public abstract String getProjectId();
 
-  public abstract String getServiceAccount();
-
-  public abstract String getPrivateKeyFile();
+  public abstract String getServiceAccountJsonKeyFile();
 }