+++ gcsio/src/test/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageFileSystemNewIntegrationTest.java
@@ -31,7 +31,7 @@ import static java.util.concurrent.TimeUnit.MINUTES;
 import static java.util.stream.Collectors.toList;
 import static org.junit.Assert.assertThrows;
 
-import com.google.api.client.auth.oauth2.Credential;
+import com.google.auth.Credentials;
 import com.google.cloud.hadoop.gcsio.integration.GoogleCloudStorageTestHelper;
 import com.google.cloud.hadoop.util.RetryHttpInitializer;
 import com.google.common.collect.ImmutableList;
@@ -64,16 +64,17 @@ public class GoogleCloudStorageFileSystemNewIntegrationTest {
 
   @BeforeClass
   public static void before() throws Throwable {
-    Credential credential =
-        checkNotNull(GoogleCloudStorageTestHelper.getCredential(), "credential must not be null");
+    Credentials credentials =
+        checkNotNull(GoogleCloudStorageTestHelper.getCredentials(), "credentials must not be null");
 
     gcsOptions = getStandardOptionBuilder().build();
     httpRequestsInitializer =
-        new RetryHttpInitializer(credential, gcsOptions.toRetryHttpInitializerOptions());
+        new RetryHttpInitializer(
+            /* delegate= */ null, credentials, gcsOptions.toRetryHttpInitializerOptions());
 
     GoogleCloudStorageFileSystem gcsfs =
         new GoogleCloudStorageFileSystem(
-            credential,
+            credentials,
             GoogleCloudStorageFileSystemOptions.builder()
                 .setBucketDeleteEnabled(true)
                 .setCloudStorageOptions(gcsOptions)
@@ -125,7 +126,7 @@ public class GoogleCloudStorageFileSystemNewIntegrationTest {
     String dirObject = getTestResource();
     URI dirObjectUri = new URI("gs://" + bucketName).resolve(dirObject);
 
-    // create directory before hand without tracking requests
+    // create directory beforehand without tracking requests
     gcsfsIHelper.mkdir(bucketName, dirObject);
     gcsFs.mkdir(dirObjectUri);
 
@@ -180,7 +181,7 @@ public class GoogleCloudStorageFileSystemNewIntegrationTest {
     String dirObject = getTestResource();
     URI dirObjectUri = new URI("gs://" + bucketName).resolve(dirObject + "/d1/");
 
-    // create directory before hand without tracking requests
+    // create directory beforehand without tracking requests
     gcsfsIHelper.mkdirs(dirObjectUri);
     gcsFs.mkdirs(dirObjectUri);