+++ util-hadoop/src/test/java/com/google/cloud/hadoop/util/AccessTokenProviderCredentialsFactoryTest.java
@@ -14,32 +14,29 @@
 
 package com.google.cloud.hadoop.util;
 
-import static com.google.cloud.hadoop.util.HadoopCredentialConfiguration.ACCESS_TOKEN_PROVIDER_IMPL_SUFFIX;
+import static com.google.cloud.hadoop.util.HadoopCredentialsConfiguration.ACCESS_TOKEN_PROVIDER_IMPL_SUFFIX;
 import static com.google.common.truth.Truth.assertThat;
 
-import com.google.api.client.auth.oauth2.Credential;
-import com.google.api.client.googleapis.auth.oauth2.GoogleCredential;
-import com.google.api.client.util.Clock;
-import com.google.cloud.hadoop.util.CredentialFromAccessTokenProviderClassFactory.GoogleCredentialWithAccessTokenProvider;
+import com.google.auth.oauth2.AccessToken;
+import com.google.auth.oauth2.GoogleCredentials;
+import com.google.cloud.hadoop.util.AccessTokenProviderCredentialsFactory.AccessTokenProviderCredentials;
 import com.google.cloud.hadoop.util.testing.TestingAccessTokenProvider;
 import com.google.common.collect.ImmutableList;
 import java.io.IOException;
-import java.security.GeneralSecurityException;
+import java.util.Date;
 import org.apache.hadoop.conf.Configuration;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
-import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 @RunWith(JUnit4.class)
-public class CredentialFromAccessTokenProviderClassFactoryTest {
+public class AccessTokenProviderCredentialsFactoryTest {
 
   private static final String TEST_PROPERTY_PREFIX = "test.prefix";
 
   private Configuration config;
-  @Mock private Clock clock;
 
   @Before
   public void setUp() {
@@ -52,25 +49,28 @@ public class CredentialFromAccessTokenProviderClassFactoryTest {
   }
 
   @Test
-  public void testCreateCredentialFromProviderClassFactory()
-      throws IOException, GeneralSecurityException {
-    Credential credential =
-        CredentialFromAccessTokenProviderClassFactory.credential(
-            config, ImmutableList.of(TEST_PROPERTY_PREFIX), ImmutableList.of());
+  public void testCreateCredentialsFromProviderClassFactory() throws IOException {
+    GoogleCredentials credentials =
+        AccessTokenProviderCredentialsFactory.credentials(
+            config, ImmutableList.of(TEST_PROPERTY_PREFIX));
 
-    assertThat(credential.getAccessToken()).isEqualTo(TestingAccessTokenProvider.FAKE_ACCESS_TOKEN);
-    assertThat(credential.getExpirationTimeMilliseconds())
-        .isEqualTo(TestingAccessTokenProvider.EXPIRATION_TIME_MILLISECONDS);
+    AccessToken accessToken = credentials.getAccessToken();
+
+    assertThat(accessToken).isNotNull();
+    assertThat(accessToken.getTokenValue()).isEqualTo(TestingAccessTokenProvider.FAKE_ACCESS_TOKEN);
+    assertThat(accessToken.getExpirationTime())
+        .isEqualTo(Date.from(TestingAccessTokenProvider.EXPIRATION_TIME));
   }
 
   @Test
-  public void testCreateCredentialFromAccessTokenProvider() {
+  public void testCreateCredentialsFromAccessTokenProvider() {
     AccessTokenProvider accessTokenProvider = new TestingAccessTokenProvider();
-    GoogleCredential credential =
-        GoogleCredentialWithAccessTokenProvider.fromAccessTokenProvider(clock, accessTokenProvider);
+    GoogleCredentials credentials = new AccessTokenProviderCredentials(accessTokenProvider);
 
-    assertThat(credential.getAccessToken()).isEqualTo(TestingAccessTokenProvider.FAKE_ACCESS_TOKEN);
-    assertThat(credential.getExpirationTimeMilliseconds())
-        .isEqualTo(TestingAccessTokenProvider.EXPIRATION_TIME_MILLISECONDS);
+    AccessToken accessToken = credentials.getAccessToken();
+    assertThat(accessToken).isNotNull();
+    assertThat(accessToken.getTokenValue()).isEqualTo(TestingAccessTokenProvider.FAKE_ACCESS_TOKEN);
+    assertThat(accessToken.getExpirationTime())
+        .isEqualTo(Date.from(TestingAccessTokenProvider.EXPIRATION_TIME));
   }
 }